some backend codes have been moved from ExpenseTracker  to ExpenseTracker.Application for security and to improve code structure
