+++ ExpenseTracker/ExpenseTracker.Apllication/Class1.cs
@@ -1,12 +1,12 @@
 ﻿using ExpenseTracker.Domain.Interfaces;
 
-namespace ExpenseTracker.Apllication
+namespace ExpenseTracker.Application
 {
     public class Class1
     {
         public Class1(ICommonRepository commonRepository)
         {
-            
+
         }
     }
 }
