+++ ExpenseTracker/ExpenseTracker.Apllication/ExpenseTracker.Application.csproj
@@ -7,7 +7,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\ExpenceTracker.Domain\ExpenseTracker.Domain.csproj" />
+	<PackageReference Include="Microsoft.AspNetCore.Http.Features" Version="5.0.17" />
+  </ItemGroup>
+
+  <ItemGroup>
+	<ProjectReference Include="..\ExpenceTracker.Domain\ExpenseTracker.Domain.csproj" />
   </ItemGroup>
 
 </Project>
