+++ ExpenseTracker/ExpenseTracker.Apllication/Extensions/DependencyInjection.cs
@@ -1,0 +1,17 @@
+﻿using ExpenseTracker.Application.Store;
+using ExpenseTracker.Application.Store.Interfaces;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace ExpenseTracker.Application.Extensions;
+
+public static class DependencyInjection
+{
+    public static IServiceCollection RegisterApplication(this IServiceCollection services, IConfiguration configuration)
+    {
+        services.AddScoped<ICategoryStore, CategoryStore>();
+        services.AddScoped<ITransferStore, TransferStore>();
+
+        return services;
+    }
+}
