+++ ExpenseTracker/ExpenseTracker.Apllication/Mappings/CategoryMappings.cs
@@ -1,7 +1,7 @@
-﻿using ExpenseTracker.Domain.Entities;
-using ExpenseTracker.ViewModels.Category;
+﻿using ExpenseTracker.Application.ViewModels.Category;
+using ExpenseTracker.Domain.Entities;
 
-namespace ExpenseTracker.Mappings;
+namespace ExpenseTracker.Application.Mappings;
 
 public static class CategoryMappings
 {
