+++ ExpenseTracker/ExpenseTracker.Apllication/Mappings/TransferMappings.cs
@@ -1,13 +1,13 @@
-﻿using ExpenseTracker.Domain.Entities;
-using ExpenseTracker.ViewModels.Transfer;
+﻿using ExpenseTracker.Application.ViewModels.Transfer;
+using ExpenseTracker.Domain.Entities;
 
-namespace ExpenseTracker.Mappings;
+namespace ExpenseTracker.Application.Mappings;
 
 public static class TransferMappings
 {
     public static TransferViewModel ToViewModel(this Transfer transfer)
     {
-        return new TransferViewModel
+        return new TransferViewModel()
         {
             Id = transfer.Id,
             Note = transfer.Note,
@@ -17,7 +17,6 @@ public static class TransferMappings
             Images = transfer.Images.Select(x => Convert.ToBase64String(x.Data)).ToList()
         };
     }
-
     public static UpdateTransferViewModel ToUpdateViewModel(this Transfer transfer)
     {
         return new UpdateTransferViewModel
