+++ ExpenseTracker/ExpenseTracker.Apllication/Store/Interfaces/ICategoryStore.cs
@@ -1,6 +1,11 @@
-﻿using ExpenseTracker.ViewModels.Category;
+﻿using ExpenseTracker.Application.ViewModels.Category;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
-namespace ExpenseTracker.Stores.Interfaces;
+namespace ExpenseTracker.Application.Store.Interfaces;
 
 public interface ICategoryStore
 {
