+++ ExpenseTracker/ExpenseTracker.Apllication/ViewModels/Account/ForgotPasswordViewModel.cs
@@ -1,0 +1,10 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace ExpenseTracker.Application.ViewModels.Account;
+
+public class ForgotPasswordViewModel
+{
+    [Required(ErrorMessage = "Email is required")]
+    [EmailAddress(ErrorMessage = "Invalid email address")]
+    public string Email { get; set; }
+}
