+++ ExpenseTracker/ExpenseTracker.Apllication/ViewModels/Category/UpdateCategoryViewModel.cs
@@ -1,7 +1,6 @@
-﻿namespace ExpenseTracker.ViewModels.Category;
+﻿namespace ExpenseTracker.Application.ViewModels.Category;
 
 public class UpdateCategoryViewModel : CreateCategoryViewModel
 {
     public int Id { get; set; }
-
 }
