+++ ExpenseTracker/ExpenseTracker.Apllication/ViewModels/Transfer/CreateTransferViewModel.cs
@@ -1,6 +1,7 @@
-﻿using System.ComponentModel.DataAnnotations;
+﻿using Microsoft.AspNetCore.Http;
+using System.ComponentModel.DataAnnotations;
 
-namespace ExpenseTracker.ViewModels.Transfer;
+namespace ExpenseTracker.Application.ViewModels.Transfer;
 
 public class CreateTransferViewModel
 {
