+++ ExpenseTracker/ExpenseTracker/Controllers/TransfersController.cs
@@ -1,5 +1,5 @@
-﻿using ExpenseTracker.Stores.Interfaces;
-using ExpenseTracker.ViewModels.Transfer;
+﻿using ExpenseTracker.Application.Store.Interfaces;
+using ExpenseTracker.Application.ViewModels.Transfer;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
 using Microsoft.EntityFrameworkCore;
