--- ExpenseTracker/ExpenseTracker/ViewModels/Account/ForgotPasswordViewModel.cs
@@ -1,11 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-
-namespace ExpenseTracker.ViewModels.Account
-{
-    public class ForgotPasswordViewModel
-    {
-        [Required(ErrorMessage = "Email is required")]
-        [EmailAddress(ErrorMessage = "Invalid email address")]
-        public string Email { get; set; }
-    }
-}
