--- ExpenseTracker/ExpenseTracker/ViewModels/Account/LoginViewModel.cs
@@ -1,18 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-
-namespace ExpenseTracker.ViewModels.Account
-{
-    public class LoginViewModel
-    {
-        [Required(ErrorMessage = "Email is required")]
-        [EmailAddress(ErrorMessage = "Invalid email address")]
-        public string Email { get; set; }
-
-        [Required(ErrorMessage = "Password is required")]
-        [DataType(DataType.Password)]
-        public string Password { get; set; }
-
-        [Display(Name = "Remember me")]
-        public bool RememberMe { get; set; }
-    }
-}
