+++ ExpenseTracker/ExpenseTracker/Views/Account/ForgotPassword.cshtml
@@ -1,4 +1,4 @@
-@model ExpenseTracker.ViewModels.Account.ForgotPasswordViewModel
+@model ExpenseTracker.Application.ViewModels.Account.ForgotPasswordViewModel
 
 @{
     ViewData["PageTitle"] = "Forgot Password";
