+++ ExpenseTracker/ExpenseTracker/Views/Account/Login.cshtml
@@ -1,4 +1,4 @@
-@model ExpenseTracker.ViewModels.Account.LoginViewModel
+@model ExpenseTracker.Application.ViewModels.Account.LoginViewModel
 
 @{
     ViewData["PageTitle"] = "Login";
