+++ ExpenseTracker/ExpenseTracker/Views/Account/Register.cshtml
@@ -1,4 +1,4 @@
-@model ExpenseTracker.ViewModels.Account.RegisterViewModel
+@model ExpenseTracker.Application.ViewModels.Account.RegisterViewModel
 
 @{
 	ViewData["PageTitle"] = "Register";
