+++ ExpenseTracker/ExpenseTracker/Views/Categories/Create.cshtml
@@ -1,4 +1,4 @@
-﻿@model ExpenseTracker.ViewModels.Category.CreateCategoryViewModel;
+﻿@model ExpenseTracker.Application.ViewModels.Category.CategoryViewModel
 
 @{
 }
