+++ ExpenseTracker/ExpenseTracker/Views/Categories/Delete.cshtml
@@ -1,4 +1,4 @@
-﻿@model ExpenseTracker.ViewModels.Category.CategoryViewModel
+﻿@model ExpenseTracker.Application.ViewModels.Category.CategoryViewModel
 
 @{
     ViewData["Title"] = "Delete";
