+++ ExpenseTracker/ExpenseTracker/Views/Categories/Details.cshtml
@@ -1,4 +1,4 @@
-﻿@model ExpenseTracker.ViewModels.Category.CategoryViewModel
+﻿@model ExpenseTracker.Application.ViewModels.Category.CategoryViewModel
 
 @{
     ViewData["Title"] = "Details";
