+++ ExpenseTracker/ExpenseTracker/Views/Categories/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@model ExpenseTracker.ViewModels.Category.UpdateCategoryViewModel;
+﻿@model ExpenseTracker.Application.ViewModels.Category.UpdateCategoryViewModel
 
 @{
 }
