+++ ExpenseTracker/ExpenseTracker/Views/Categories/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<ExpenseTracker.ViewModels.Category.CategoryViewModel>
+﻿@model IEnumerable<ExpenseTracker.Application.ViewModels.Category.CategoryViewModel>
 
 @{
     ViewData["PageTitle"] = "Categories";
