+++ ExpenseTracker/ExpenseTracker/Views/Transfers/Create.cshtml
@@ -1,5 +1,5 @@
-﻿@using ExpenseTracker.Domain.Enums
-@model ExpenseTracker.ViewModels.Transfer.CreateTransferViewModel
+﻿@using ExpenseTracker.Domain.Enums;
+@model ExpenseTracker.Application.ViewModels.Transfer.CreateTransferViewModel
 
 @{
     ViewData["PageTitle"] = "Create a new transaction";
