+++ ExpenseTracker/ExpenseTracker/Views/Transfers/Details.cshtml
@@ -1,4 +1,4 @@
-﻿@model ExpenseTracker.ViewModels.Transfer.TransferViewModel
+﻿@model ExpenseTracker.Application.ViewModels.Transfer.TransferViewModel
 
 @{
     ViewData["Title"] = "Details";
