+++ ExpenseTracker/ExpenseTracker/Views/Transfers/Edit.cshtml
@@ -1,6 +1,6 @@
 ﻿@using ExpenseTracker.Domain.Enums
 
-@model ExpenseTracker.ViewModels.Transfer.UpdateTransferViewModel
+@model ExpenseTracker.Application.ViewModels.Transfer.UpdateTransferViewModel
 
 @{
     ViewData["Title"] = "Edit";
