+++ ExpenseTracker/ExpenseTracker/Views/Transfers/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<ExpenseTracker.ViewModels.Transfer.TransferViewModel>
+﻿@model IEnumerable<ExpenseTracker.Application.ViewModels.Transfer.TransferViewModel>
 
 @{
     ViewData["PageTitle"] = "Transactions";
