+++ ExpenseTracker/ExpenseTracker/appsettings.json
@@ -1,0 +1,20 @@
+﻿{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "ExpenseTrackerDbContextConnection": "Data Source=desktop-vvif9hi\\ramadan;Initial Catalog=ExpenseTrackerG5;Integrated Security=True;Pooling=False;Encrypt=True;Trust Server Certificate=True"
+  },
+  "MailSettings": {
+    "From": "shukhratovich75@gmail.com",
+    "SmtpServer": "smtp.gmail.com",
+    "Port": 465,
+    "Username": "shukhratovich75@gmail.com",
+    "Password": "jjej qecs qmcd lenq"
+  },
+  "SyncfusionKey": "Ngo9BigBOggjHTQxAR8/V1NCaF5cXmZCf1FpRmJGdld5fUVHYVZUTXxaS00DNHVRdkdnWXdedHRQRGRZWEV+WkQ="
+}
